fix warnings, navbar style change, job filtering!
ReactJS uses autobind, so remove the .bind(this) parameters, replacing them with
null when needed.

After adding a search bar to the top right of the screen, it looked out of place
floating off to the right. Change the tabs to a nav bar and add the search box to
the right.

Server side (with regular expressions) filtering added. It was the
easiest/laziest way I thought of, since it's just a simple injected filter.

A little bit of padding between each job post to make it feel a bit easier to
read each individual post.

General modernization of the reactjs code.